Configuração dos endpoints e service ClientePJ

+++ src/main/java/sousa/banco/service/ClientePJService.java
@@ -1,9 +1,15 @@
 package sousa.banco.service;
 
+import io.quarkus.cache.CacheInvalidateAll;
+import io.quarkus.cache.CacheResult;
+import jakarta.annotation.security.RolesAllowed;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.transaction.Transactional;
+import org.eclipse.microprofile.openapi.annotations.security.SecurityRequirement;
 import sousa.banco.dto.ClientePJDTO;
 import sousa.banco.entity.*;
+import sousa.banco.exception.ConflictException;
+import sousa.banco.exception.NotFoundException;
 import sousa.banco.mapper.*;
 import sousa.banco.repository.ClientePJRepository;
 
@@ -19,7 +25,16 @@ public class ClientePJService {
     }
 
     @Transactional
+    @RolesAllowed({"ROLE_ADMIN", "ROLE_USER"})
+    @SecurityRequirement(name = "Keycloak")
+    @CacheInvalidateAll(cacheName = "clientePJPorId")
+    @CacheInvalidateAll(cacheName = "clientesPJTodos")
+    @CacheInvalidateAll(cacheName = "clientePJPorCnpj")
     public void criaClientePJ(ClientePJDTO clientePJDTO) {
+        ClientePJ clienteExistente = clientePJRepository.buscaPorCnpj(clientePJDTO.cnpj());
+        if (clienteExistente != null) {
+            throw new ConflictException("ClientePJ com CNPJ " + clientePJDTO.cnpj() + " já existe.");
+        }
 
         ClientePJ clientePJ = ClientePJMapper.toEntity(clientePJDTO);
 
@@ -55,10 +70,15 @@ public class ClientePJService {
     }
 
     @Transactional
+    @RolesAllowed({"ROLE_ADMIN", "ROLE_USER"})
+    @SecurityRequirement(name = "Keycloak")
+    @CacheInvalidateAll(cacheName = "clientePJPorId")
+    @CacheInvalidateAll(cacheName = "clientesPJTodos")
+    @CacheInvalidateAll(cacheName = "clientePJPorCnpj")
     public ClientePJDTO atualizaClientePJ(Long id, ClientePJDTO clientePJDTO) {
         ClientePJ clientePJ = clientePJRepository.findById(id);
         if (clientePJ == null) {
-            throw new RuntimeException("Cliente PJ não encontrado");
+            throw new NotFoundException("Cliente PJ com ID " + id + " não encontrado");
         }
 
         clientePJ.setRazaoSocial(clientePJDTO.razaoSocial());
@@ -88,7 +108,7 @@ public class ClientePJService {
         }
 
         if (clientePJDTO.socios() != null) {
-            //clientePJ.getSocios().clear();
+            clientePJ.getSocios().clear();
             clientePJDTO.socios().forEach(sDto -> {
                 ParticipacaoSocietaria socio = ParticipacaoSocietariaMapper.toEntity(sDto);
                 clientePJ.addSocio(socio);
@@ -99,24 +119,54 @@ public class ClientePJService {
     }
 
     @Transactional
+    @RolesAllowed({"ROLE_ADMIN"})
+    @SecurityRequirement(name = "Keycloak")
+    @CacheInvalidateAll(cacheName = "clientePJPorId")
+    @CacheInvalidateAll(cacheName = "clientesPJTodos")
+    @CacheInvalidateAll(cacheName = "clientePJPorCnpj")
     public boolean deletaClientePJ(Long id) {
+        var ClientePJ =  clientePJRepository.findById(id);
+        if (ClientePJ == null) {
+            throw new NotFoundException("Cliente PJ com ID " + id + " não encontrado");
+        }
         clientePJRepository.deleteById(id);
         return true;
     }
 
+    @RolesAllowed({"ROLE_ADMIN", "ROLE_USER"})
+    @SecurityRequirement(name = "Keycloak")
+    @CacheResult(cacheName = "clientePJPorId")
     public ClientePJDTO buscaClientePJPorId(Long id) {
         ClientePJ clientePJ = clientePJRepository.findById(id);
-        if (clientePJ != null) {
-            return ClientePJMapper.toDTO(clientePJ);
-        } else {
-            throw new RuntimeException("Cliente PJ não encontrado");
+        if (clientePJ == null) {
+            throw new NotFoundException("Cliente PJ com ID " + id + " não encontrado");
         }
+        return ClientePJMapper.toDTO(clientePJ);
     }
 
+    @RolesAllowed({"ROLE_ADMIN", "ROLE_USER"})
+    @SecurityRequirement(name = "Keycloak")
+    @CacheResult(cacheName = "clientePJTodos")
     public List<ClientePJDTO> buscaTodosClientesPJ() {
-        return clientePJRepository.listAll()
-                .stream()
+        List<ClientePJ> clientesPJ = clientePJRepository.listAll();
+
+        if (clientesPJ.isEmpty()) {
+            throw new NotFoundException("Nenhum Cliente PJ encontrado.");
+        }
+
+        return clientesPJ.stream()
                 .map(ClientePJMapper::toDTO)
                 .toList();
+    }
+
+    @RolesAllowed({"ROLE_ADMIN", "ROLE_USER"})
+    @SecurityRequirement(name = "Keycloak")
+    @CacheResult(cacheName = "clientePJPorCnpj")
+    public ClientePJDTO buscaClientePJPorCnpj(String cnpj) {
+        ClientePJ clientePJ = clientePJRepository.buscaPorCnpj(cnpj);
+        if (clientePJ == null) {
+            throw new NotFoundException("Cliente PJ com CNPJ " + cnpj + " não encontrado.");
+        }
+        return ClientePJMapper.toDTO(clientePJ);
     }
 }
